Update bill metadata when adding new actions (#12)
Bill info (title/description) can change over time and the API does not offer a good way to query for these changes. In this PR we start checking for updated bill information every time we look for new actions and scheduled actions. (The latter might be overkill, but it's at worst overkill.)